refactor: ENUM 클래스 패키지 변경
- code 패키지로 이동
