+++ wecare-backend/wecare/src/main/java/com/example/wecare/member/code/Gender.java
@@ -1,0 +1,5 @@
+package com.example.wecare.member.code;
+
+public enum Gender {
+    MALE, FEMALE
+}
