+++ wecare-backend/wecare/src/main/java/com/example/wecare/member/code/Role.java
@@ -1,4 +1,4 @@
-package com.example.wecare.member.domain;
+package com.example.wecare.member.code;
 
 public enum Role {
     GUARDIAN, DEPENDENT
