--- wecare-backend/wecare/src/main/java/com/example/wecare/member/domain/Gender.java
@@ -1,5 +1,0 @@
-package com.example.wecare.member.domain;
-
-public enum Gender {
-    MALE, FEMALE
-}
