+++ wecare-backend/wecare/src/main/java/com/example/wecare/routine/code/RepeatDay.java
@@ -1,4 +1,4 @@
-package com.example.wecare.routine.domain;
+package com.example.wecare.routine.code;
 
 public enum RepeatDay {
     MON, TUE, WED, THU, FRI, SAT, SUN, DAILY
