+++ wecare-backend/wecare/src/main/java/com/example/wecare/routine/code/RoutineType.java
@@ -1,0 +1,5 @@
+package com.example.wecare.routine.code;
+
+public enum RoutineType {
+    ACTIVITY, CUSTOM, MEDICATION, SUPPLEMENT, MEAL
+}
