+++ wecare-backend/wecare/src/main/java/com/example/wecare/routine/code/SoundType.java
@@ -1,0 +1,5 @@
+package com.example.wecare.routine.code;
+
+public enum SoundType {
+    DEFAULT_SOUND, SILENT, VIBRATION, VOICE_MESSAGE
+}
