--- wecare-backend/wecare/src/main/java/com/example/wecare/routine/domain/RoutineType.java
@@ -1,5 +1,0 @@
-package com.example.wecare.routine.domain;
-
-public enum RoutineType {
-    MEDICATION, SUPPLEMENT, ACTIVITY, CUSTOM, MEAL
-}
