--- wecare-backend/wecare/src/main/java/com/example/wecare/routine/domain/SoundType.java
@@ -1,8 +1,0 @@
-package com.example.wecare.routine.domain;
-
-public enum SoundType {
-    DEFAULT_SOUND, // 기본 알림음
-    VOICE_MESSAGE, // 음성 메시지 (녹음)
-    VIBRATION,     // 진동
-    SILENT         // 무음
-}
